% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_k}
\alias{select_k}
\title{Select Number of Components}
\usage{
select_k(
  Y,
  X,
  k,
  rho = 0,
  tol = 1e-10,
  maxit = 1000,
  center = TRUE,
  scale = FALSE,
  quiet = TRUE,
  L,
  m = 2
)
}
\arguments{
\item{Y}{n x r matrix of responses}

\item{X}{n x p matrix of predictors}

\item{k}{Vector of numbers of principal components to consider}

\item{rho}{Ridge penalty on alpha in representation beta = L alpha}

\item{tol}{Tolerance for L-BFGS-B on profile likelihood}

\item{maxit}{Maximum number of iterations of L-BFGS-B algorithm}

\item{center}{If TRUE, responses and predictors are centered by their sample mean}

\item{scale}{If TRUE, divide each column in X by its sample standard deviation}

\item{quiet}{If FALSE, print information from L-BFGS-B algorithm}

\item{L}{Starting value in L-BFGS-B for the Cholesky root
in the decomposition Sigma_X = tau (I + LL')}

\item{m}{Penalty in information criterion
- 2 * log-likelihood + m * n_params}
}
\value{
List with selected k, information criterion for all k, and model fit
with selcted k.
}
\description{
This function selects the number of components for jlpcr
}
