% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{jlpcr}
\alias{jlpcr}
\title{Joint Likelihood Principal Components Regression}
\usage{
jlpcr(
  Y,
  X,
  k,
  rho = 0,
  tol = 1e-10,
  maxit = 1000,
  center = TRUE,
  scale = FALSE,
  quiet = TRUE,
  L,
  m = 2
)
}
\arguments{
\item{Y}{n x r matrix of responses}

\item{X}{n x p matrix of predictors}

\item{k}{Number of principal components to use}

\item{rho}{Ridge penalty on alpha in representation beta = L alpha}

\item{tol}{Tolerance for L-BFGS-B on profile likelihood}

\item{maxit}{Maximum number of iterations of L-BFGS-B algorithm}

\item{center}{If TRUE, responses and predictors are centered by their sample mean}

\item{scale}{If TRUE, divide each column in X by its sample standard deviation}

\item{quiet}{If FALSE, print information from L-BFGS-B algorithm}

\item{L}{Starting value in L-BFGS-B for the Cholesky root
in the decomposition Sigma_X = tau (I + LL')}

\item{m}{Penalty in information criterion
- 2 * log-likelihood + m * n_params; can be a vector}
}
\value{
List with estimates:
         beta (regression coefficient),
         Sigma (response covariance matrix),
         Sigma_X (predictor covariance matrix),
         L (Cholesky root in decomposition Sigma_X = tau[I + LL']),
         tau (smallest eigenvalue of Sigma_X; has multiplicity p - k), and
         IC (information criterion)
}
\description{
\code{jlpcr} fits a principal components regression by maximizing the joint likelihood 
 of multivariate normal responses and predictors.
}
\details{
This is the only function exported from the package with the same name. 
The likelihood maximized is that for n independent observations 
from a normal multivariate response linear regression model where the column space
of the p-by-r coefficient matrix is spanned by the k leading eigenvectors (corresponding to
the largest eigenvalues) of the predictors' covariance matrix.
This covariance matrix is assumed to be spiked, meaning its
smallest eigenvalue has multiplicity p - k.
}
