% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{jlpcr}
\alias{jlpcr}
\title{Fit Principal Components Regression using the jointly normal likelihood}
\usage{
jlpcr(Y, X, k, rho = 0, tol = 1e-10, maxit = 1000, center = TRUE,
  quiet = TRUE, L)
}
\arguments{
\item{Y}{An n x r matrix of responses.}

\item{X}{An n x p matrix of predictors}

\item{k}{Number of principal components to use}

\item{rho}{Ridge penalty on alpha in representation beta = L alpha}

\item{tol}{The tolerance for L-BFGS-G algorithm used on profile likelihood}

\item{maxit}{The maximum number of iterations of L-BFGS-G algorithm}

\item{center}{If TRUE, responses and predictors are centered by their sample mean}

\item{quiet}{If FALSE, print print information from L-BFGS-G algorithm}

\item{L}{Starting value for the Cholesky root in the decomposition
Sigma_X = tau (I_p + LL^T)}
}
\value{
List with estimates of beta (regression coefficient),
        Sigma (response covariance matrix),
        Sigma_X (predictor covariance matrix),
        L (Cholesky root in decomposition of Sigma_X), and 
        tau (smallest eigenvalue of Sigma_X; has multiplicity p - k)
}
\description{
Fit Principal Components Regression using the jointly normal likelihood
}
